Features/authorization

+++ controllers/card_controller.py
@@ -6,6 +6,7 @@ from flask_jwt_extended import jwt_required, get_jwt_identity
 from init import db
 from models.card import Card, card_schema, cards_schema
 from controllers.comment_controller import comments_bp
+from utils import authorise_as_admin
 
 cards_bp = Blueprint("cards", __name__, url_prefix="/cards")
 cards_bp.register_blueprint(comments_bp)
@@ -50,10 +51,15 @@ def create_card():
 @cards_bp.route("/<int:card_id>", methods=["DELETE"])
 @jwt_required()
 def delete_card(card_id):
+    # check admin status
+    
     stmt = db.select(Card).filter_by(id=card_id)
     card = db.session.scalar(stmt)
 
     if card:
+        is_admin = authorise_as_admin()
+        if not is_admin or str(card.user_id) != get_jwt_identity():
+            return {"error": "User is not authorised to perform this action"}, 403
         db.session.delete(card)
         db.session.commit()
         return {"message": f"Card '{card.title}' deleted successfully"}
@@ -69,6 +75,9 @@ def update_card(card_id):
     card = db.session.scalar(stmt)
 
     if card:
+        # if the user is not the owner of the card
+        if str(card.user_id) != get_jwt_identity():
+            return {"error": "You are not the owner of the card"}, 403
         
         card.title = body_data.get("title") or card.title,
         card.description=body_data.get("description") or card.description,
@@ -80,3 +89,4 @@ def update_card(card_id):
 
     else:
         return {"error": f"Card with id {card_id} not found"}, 404
+
